+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+## Ormolu action v17
+
+* Uses Ormolu 0.8.0.0 as the latest version.
+
+## Ormolu action v16
+
+* Uses Ormolu 0.7.7.0 as the latest version. On MacOS runners aarch64
+  version of Ormolu will be used.
+
+## Ormolu action v15
+
+* Uses Ormolu 0.7.4.0 as the latest version.
+
+## Ormolu action v14
+
+* Added the `mode` input. [PR
+  29](https://github.com/haskell-actions/run-ormolu/pull/29).
+
+## Ormolu action v13
+
+* Uses Ormolu 0.7.1.0 as the latest version.
+
 ## Ormolu action v12
 
 * Uses Ormolu 0.7.0.0 as the latest version.